heroku setup fixes

+++ db/migrations/20190507090924_create_site_table.js
@@ -1,16 +1,16 @@
-exports.up = function(knex, Promise) {
+exports.up = function (knex, Promise) {
   return knex.schema.createTable('sites', siteTable => {
     siteTable.increments('site_id').primary();
     siteTable.string('name');
-    siteTable.float('latitude_min', null);
-    siteTable.float('longitude_min', null);
-    siteTable.float('altitude_min', null);
-    siteTable.float('latitude_max', null);
-    siteTable.float('longitude_max', null);
-    siteTable.float('altitude_max', null);
+    siteTable.decimal('latitude_min', null);
+    siteTable.decimal('longitude_min', null);
+    siteTable.decimal('altitude_min', null);
+    siteTable.decimal('latitude_max', null);
+    siteTable.decimal('longitude_max', null);
+    siteTable.decimal('altitude_max', null);
   });
 };
 
-exports.down = function(knex, Promise) {
+exports.down = function (knex, Promise) {
   return knex.schema.dropTable('sites');
 };
