+++ db/migrations/20190507090930_create_pin_table.js
@@ -1,4 +1,4 @@
-exports.up = function(knex, Promise) {
+exports.up = function (knex, Promise) {
   return knex.schema.createTable('pins', pinTable => {
     pinTable.increments('pin_id').primary();
     pinTable
@@ -10,14 +10,14 @@ exports.up = function(knex, Promise) {
       .references('sites.site_id')
       .onDelete('CASCADE');
     pinTable.timestamp('timestamp').defaultTo(knex.fn.now());
-    pinTable.float('latitude', null);
-    pinTable.float('longitude', null);
-    pinTable.float('altitude', null);
+    pinTable.decimal('latitude', null);
+    pinTable.decimal('longitude', null);
+    pinTable.decimal('altitude', null);
     pinTable.string('photo_url', 1000);
     pinTable.text('note');
   });
 };
 
-exports.down = function(knex, Promise) {
+exports.down = function (knex, Promise) {
   return knex.schema.dropTable('pins');
 };
